move `SingleResidueVariant` into residue.py
