+++ deeprank2/features/conservation.py
@@ -1,13 +1,12 @@
 from typing import Optional
 
 import numpy as np
-from deeprank2.domain.aminoacidlist import amino_acids
-from deeprank2.molstruct.atom import Atom
-from deeprank2.molstruct.residue import Residue
-from deeprank2.molstruct.variant import SingleResidueVariant
-from deeprank2.utils.graph import Graph
 
 from deeprank2.domain import nodestorage as Nfeat
+from deeprank2.domain.aminoacidlist import amino_acids
+from deeprank2.molstruct.atom import Atom
+from deeprank2.molstruct.residue import Residue, SingleResidueVariant
+from deeprank2.utils.graph import Graph
 
 
 def add_features( # pylint: disable=unused-argument
