+++ deeprank2/features/surfacearea.py
@@ -3,12 +3,11 @@ from typing import Optional
 
 import freesasa
 import numpy as np
-from deeprank2.molstruct.atom import Atom
-from deeprank2.molstruct.residue import Residue
-from deeprank2.molstruct.variant import SingleResidueVariant
-from deeprank2.utils.graph import Graph
 
 from deeprank2.domain import nodestorage as Nfeat
+from deeprank2.molstruct.atom import Atom
+from deeprank2.molstruct.residue import Residue, SingleResidueVariant
+from deeprank2.utils.graph import Graph
 
 # pylint: disable=c-extension-no-member
